requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.0
channels==4.1.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==4.2.13
django-cors-headers==4.4.0
django-filter==24.3
django-leaflet==0.30.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.0
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.35.1
rpds-py==0.18.1
service-identity==24.1.0
setuptools==74.0.0
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.0.3
